+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* Avoid spurious frozen string literal deprecation warning on Ruby 3.4.0-preview2+ (jeremyevans)
+
 ## 2.4.0 (2024-06-27)
 
 * Support commonmarker 1.0+ API (unasuke) (#10)